+++ Sierra/src/Rendering/Platform/Vulkan/VulkanBuffer.cpp
@@ -32,7 +32,7 @@ namespace Sierra
 
         // Create and allocate buffer
         const VkResult result = vmaCreateBuffer(device.GetMemoryAllocator(), &bufferCreateInfo, &allocationCreateInfo, &buffer, &allocation, nullptr);
-        SR_ERROR_IF(result != VK_SUCCESS, "[Vulkan]: Failed to create buffer [{0}]! Error code: {1}.", GetName(), result);
+        SR_ERROR_IF(result != VK_SUCCESS, "[Vulkan]: Failed to create buffer [{0}]! Error code: {1}.", GetName(), static_cast<int32>(result));
         device.SetObjectName(buffer, VK_OBJECT_TYPE_BUFFER, GetName());
 
         // Map and reset memory if CPU-visible
