+++ Sierra/src/Rendering/Platform/Vulkan/VulkanShader.cpp
@@ -23,7 +23,7 @@ namespace Sierra
 
         // Create shader module
         const VkResult result = device.GetFunctionTable().vkCreateShaderModule(device.GetLogicalDevice(), &shaderModuleCreateInfo, nullptr, &shaderModule);
-        SR_ERROR_IF(result != VK_SUCCESS, "[Vulkan]: Could not create shader module for shader [{0}]! Error code: {1}.", GetName(), result);
+        SR_ERROR_IF(result != VK_SUCCESS, "[Vulkan]: Could not create shader module for shader [{0}]! Error code: {1}.", GetName(), static_cast<int32>(result));
 
         // Set object name
         device.SetObjectName(shaderModule, VK_OBJECT_TYPE_SHADER_MODULE, GetName());
